exercise 7.10/11: redux step2 and step3
Connect the blog reducer to replace state held in App.

This still leaves the callback handling in App, partially because I'm
too lazy to change the tests for the components that would need to
switch to using the store, and partially because there's still a need to
control the toggling of the creation form from high in the tree.